+++ cmd/entire/cli/checkpoint/committed.go
@@ -1587,7 +1587,7 @@ func (s *GitStore) replaceSkillEvents(skillEvents []agent.SkillEvent, sessionPat
 // computed once across multiple checkpoints.
 func (s *GitStore) replaceTranscript(ctx context.Context, transcript redact.RedactedBytes, agentType types.AgentType, precomputed *PrecomputedTranscriptBlobs, sessionPath string, entries map[string]object.TreeEntry) error {
 	// Ignore precompute if invariants are violated — fall back to fresh chunking.
-	if precomputed != nil && !precomputed.isUsable() {
+	if precomputed != nil && !precomputed.IsUsable() {
 		precomputed = nil
 	}
