+++ cmd/entire/cli/checkpoint/committed_reader_resolve.go
@@ -2,84 +2,13 @@ package checkpoint
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/entireio/cli/cmd/entire/cli/checkpoint/id"
 )
 
-// CommittedReader provides read access to committed checkpoint data.
-type CommittedReader interface {
-	ReadCommitted(ctx context.Context, checkpointID id.CheckpointID) (*CheckpointSummary, error)
-	ReadSessionContent(ctx context.Context, checkpointID id.CheckpointID, sessionIndex int) (*SessionContent, error)
-}
-
-// CommittedListReader provides read and list access to committed checkpoint data.
-type CommittedListReader interface {
-	CommittedReader
-	ListCommitted(ctx context.Context) ([]CommittedInfo, error)
-	ReadSessionMetadata(ctx context.Context, checkpointID id.CheckpointID, sessionIndex int) (*CommittedMetadata, error)
-	ReadSessionPrompts(ctx context.Context, checkpointID id.CheckpointID, sessionIndex int) (string, error)
-}
-
-// CommittedStore provides the production committed checkpoint storage surface.
-// Writes go through the unified Writer.Write(ctx, WriteRequest); the concrete
-// per-operation methods (WriteCommitted/UpdateCommitted/...) remain on GitStore
-// as the implementation Write dispatches to.
-type CommittedStore interface {
-	CommittedListReader
-	ReadSessionMetadataAndPrompts(ctx context.Context, checkpointID id.CheckpointID, sessionIndex int) (*SessionContent, error)
-	Writer
-}
-
-// AuthorReader provides optional checkpoint author lookup.
+// AuthorReader provides optional checkpoint author lookup. It stays in the
+// implementation package because GetCheckpointAuthor is a git-log operation
+// and Author is an implementation type, not part of the storage contract.
 type AuthorReader interface {
 	GetCheckpointAuthor(ctx context.Context, checkpointID id.CheckpointID) (Author, error)
-}
-
-// ReadCommittedCheckpoint reads a committed checkpoint summary and normalizes
-// a nil store response into ErrCheckpointNotFound.
-func ReadCommittedCheckpoint(ctx context.Context, reader CommittedReader, checkpointID id.CheckpointID) (*CheckpointSummary, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, err //nolint:wrapcheck // Propagating context cancellation
-	}
-
-	summary, err := reader.ReadCommitted(ctx, checkpointID)
-	if err != nil {
-		return nil, fmt.Errorf("read committed checkpoint: %w", err)
-	}
-	if summary == nil {
-		return nil, ErrCheckpointNotFound
-	}
-	return summary, nil
-}
-
-// ReadLatestSessionContent reads the latest session from an already-resolved
-// committed reader and summary.
-func ReadLatestSessionContent(ctx context.Context, reader CommittedReader, checkpointID id.CheckpointID, summary *CheckpointSummary) (*SessionContent, error) {
-	if summary == nil || len(summary.Sessions) == 0 {
-		return nil, ErrCheckpointNotFound
-	}
-	latestIndex := len(summary.Sessions) - 1
-	content, err := reader.ReadSessionContent(ctx, checkpointID, latestIndex)
-	if err != nil {
-		return nil, fmt.Errorf("read session %d content: %w", latestIndex, err)
-	}
-	return content, nil
-}
-
-func ReadRawSessionLogForCheckpoint(ctx context.Context, reader CommittedReader, checkpointID id.CheckpointID) ([]byte, string, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, "", err //nolint:wrapcheck // Propagating context cancellation
-	}
-
-	summary, err := ReadCommittedCheckpoint(ctx, reader, checkpointID)
-	if err != nil {
-		return nil, "", err
-	}
-
-	content, err := ReadLatestSessionContent(ctx, reader, checkpointID, summary)
-	if err != nil {
-		return nil, "", err
-	}
-	return content.Transcript, content.Metadata.SessionID, nil
 }
